move context to seperate folder
